requirements.txt
annotated-types==0.6.0
anyio==4.2.0
fastapi==0.109.0
greenlet==3.0.3
idna==3.6
pydantic==2.5.3
pydantic_core==2.14.6
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
jose==1.0.0
uvicorn~=0.27.0
fastapi-mqtt==2.1.0
python-dotenv~=1.0.1